Check SSH tunneling
FC dbus service runs a separate process to create the SSH tunnel.

1) The current implementation doesn't check the result of such tunneling
and thus, the failure will be deferred to the actual using of
the tunnel either by SPICE or by notifying channel.

The essential thing for creating a tunnel is 'ExitOnForwardFailure' SSH
option:

> Specifies whether ssh(1) should terminate the connection if it cannot set up
> all requested dynamic, tunnel, local, and remote port forwardings, (e.g. if
> either end is unable to bind and listen on a specified port).

The result of starting tunneling should be checked as well.

2) Another current drawback is not checking the PID of the tunnel
process before killing. Of course, some random process with the
same ID may be reaped.

With the help of SSH ControlMaster feature this could be achieved:
a) create tunnel within the master connection
b) then kill it using control socket.

Fixes: https://github.com/fleet-commander/fc-admin/issues/238
Signed-off-by: Stanislav Levin <slev@altlinux.org>