libvirt: Don't set 'connected' for virtio channel
According to the docs [0]:

> Since 1.2.11 the active XML for a virtio channel may contain an
optional state attribute that reflects whether a process in the
guest is active on the channel. This is an output-only attribute.
Possible values for the state attribute are connected and
disconnected.

Fortunately, the write operation is noop for now, but this could be
changed in future.

[0]: https://libvirt.org/formatdomain.html

Fixes: https://github.com/fleet-commander/fc-admin/issues/238
Signed-off-by: Stanislav Levin <slev@altlinux.org>