tests: Rename Python modules to conform Python Identifiers
According to Python docs:
```
Within the ASCII range (U+0001..U+007F), the valid characters for
identifiers are the same as in Python 2.x: the uppercase and lowercase
letters A through Z, the underscore _ and, except for the first
character, the digits 0 through 9.
```

With this change tests modules can be imported without any workarounds.

Signed-off-by: Stanislav Levin <slev@altlinux.org>